fix dashboard and add filtering by store in financial hub
